fix:서버바꿀때 새로운 api요청 안가는 문제 해결
